# openapi_watson.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Climate Cognitive System API",
    "description": "IoT Cognitive backend for climate sensor data ingestion and analysis.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/sensors/": {
      "get": {
        "tags": [
          "sensors"
        ],
        "summary": "List Readings",
        "operationId": "list_readings_api_v1_sensors__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "sensor_id",
            "in": "query",
            "description": "ID del sensor o aula a consultar",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "tags": [
          "sensors"
        ],
        "summary": "Ingest Reading",
        "operationId": "ingest_reading_api_v1_sensors__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorReading"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/admin/setup-rooms": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Setup Room",
        "operationId": "setup_room_api_v1_admin_setup_rooms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat",
        "operationId": "chat_api_v1_chat__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/health": {
      "get": {
        "summary": "Health Check",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChatRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "nullable": true
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "response": {
            "type": "string",
            "title": "Response"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "response",
          "session_id"
        ],
        "title": "ChatResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RoomIn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "max_capacity": {
            "type": "integer",
            "title": "Max Capacity"
          },
          "target_temp": {
            "type": "number",
            "title": "Target Temp"
          },
          "sensor_id": {
            "type": "string",
            "title": "Sensor Id",
            "nullable": true
          },
          "schedules": {
            "items": {
              "$ref": "#/components/schemas/ScheduleIn"
            },
            "type": "array",
            "title": "Schedules",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name",
          "max_capacity",
          "target_temp"
        ],
        "title": "RoomIn"
      },
      "ScheduleIn": {
        "properties": {
          "day_of_week": {
            "type": "integer",
            "title": "Day Of Week"
          },
          "start_time": {
            "type": "string",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "title": "End Time"
          },
          "expected_people": {
            "type": "integer",
            "title": "Expected People"
          }
        },
        "type": "object",
        "required": [
          "day_of_week",
          "start_time",
          "end_time",
          "expected_people"
        ],
        "title": "ScheduleIn"
      },
      "SensorReading": {
        "properties": {
          "sensor_id": {
            "type": "string",
            "title": "Sensor Id"
          },
          "temperature": {
            "type": "number",
            "title": "Temperature",
            "description": "Temperature in Celsius"
          },
          "humidity": {
            "type": "number",
            "title": "Humidity",
            "description": "Relative humidity percentage"
          },
          "co2_ppm": {
            "type": "number",
            "title": "Co2 Ppm",
            "description": "CO2 concentration in ppm",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "sensor_id",
          "temperature",
          "humidity"
        ],
        "title": "SensorReading"
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "servers": [
    {
      "url": "https://visible-hurling-scribble.ngrok-free.dev"
    }
  ]
}